[DOM-368] Refactor loading overlay
