+++ Domostroy/Domostroy/Flows/AdminPanel/UsersAdmin/Presenter/UsersAdminPresenter.swift
@@ -155,14 +155,14 @@ private extension UsersAdminPresenter {
     }
 
     func setBanned(userId: Int, value: Bool, completion: ((Bool) -> Void)?) {
-        let loading = DLoadingOverlay.shared.show()
+        let loading = LoadingOverlayPresenter.shared.show()
         loading.cancellable.store(in: &cancellables)
         self.adminService?.setUserBan(
             id: userId,
             value: value
         ).sink(
             receiveCompletion: { _ in
-                DLoadingOverlay.shared.hide(id: loading.id)
+                LoadingOverlayPresenter.shared.hide(id: loading.id)
             },
             receiveValue: { result in
                 switch result {
@@ -177,13 +177,13 @@ private extension UsersAdminPresenter {
     }
 
     func delete(userId: Int, completion: ((Bool) -> Void)?) {
-        let loading = DLoadingOverlay.shared.show()
+        let loading = LoadingOverlayPresenter.shared.show()
         loading.cancellable.store(in: &cancellables)
         self.adminService?.deleteUser(
             id: userId
         ).sink(
             receiveCompletion: { _ in
-                DLoadingOverlay.shared.hide(id: loading.id)
+                LoadingOverlayPresenter.shared.hide(id: loading.id)
             },
             receiveValue: { [weak self] result in
                 switch result {
