+++ Domostroy/Domostroy/Flows/OfferDetails/OfferDetails/Presenter/OfferDetailsPresenter.swift
@@ -251,13 +251,13 @@ private extension OfferDetailsPresenter {
     }
 
     func banOffer(id: Int, reason: String?, completion: EmptyClosure?, handleResult: ((NodeResult<Void>) -> Void)?) {
-        let loading = DLoadingOverlay.shared.show()
+        let loading = LoadingOverlayPresenter.shared.show()
         loading.cancellable.store(in: &cancellables)
         adminService?.banOffer(
             banOfferEntity: .init(offerId: id, banReason: reason)
         ).sink(
             receiveCompletion: { _ in
-                DLoadingOverlay.shared.hide(id: loading.id)
+                LoadingOverlayPresenter.shared.hide(id: loading.id)
                 completion?()
             },
             receiveValue: { handleResult?($0) }
@@ -265,13 +265,13 @@ private extension OfferDetailsPresenter {
     }
 
     func unbanOffer(id: Int, completion: EmptyClosure?, handleResult: ((NodeResult<Void>) -> Void)?) {
-        let loading = DLoadingOverlay.shared.show()
+        let loading = LoadingOverlayPresenter.shared.show()
         loading.cancellable.store(in: &cancellables)
         adminService?.unbanOffer(
             id: id
         ).sink(
             receiveCompletion: { _ in
-                DLoadingOverlay.shared.hide(id: loading.id)
+                LoadingOverlayPresenter.shared.hide(id: loading.id)
                 completion?()
             },
             receiveValue: { handleResult?($0) }
