+++ Domostroy/Domostroy/Flows/OfferDetails/UserProfile/Presenter/UserProfilePresenter.swift
@@ -293,14 +293,14 @@ private extension UserProfilePresenter {
         guard let userId else {
             return
         }
-        let loading = DLoadingOverlay.shared.show()
+        let loading = LoadingOverlayPresenter.shared.show()
         loading.cancellable.store(in: &cancellables)
         adminService?.setUserBan(
             id: userId,
             value: banned
         ).sink(
             receiveCompletion: { _ in
-                DLoadingOverlay.shared.hide(id: loading.id)
+                LoadingOverlayPresenter.shared.hide(id: loading.id)
             },
             receiveValue: { [weak self] result in
                 switch result {
