+++ Domostroy/Domostroy/Flows/Requests/OutgoingRequestDetails/Presenter/OutgoingRequestDetailsPresenter.swift
@@ -115,13 +115,13 @@ private extension OutgoingRequestDetailsPresenter {
         guard let requestId else {
             return
         }
-        let loading = DLoadingOverlay.shared.show()
+        let loading = LoadingOverlayPresenter.shared.show()
         loading.cancellable.store(in: &cancellables)
         rentService?.deleteRentRequest(
             id: requestId
         ).sink(
             receiveCompletion: { _ in
-                DLoadingOverlay.shared.hide(id: loading.id)
+                LoadingOverlayPresenter.shared.hide(id: loading.id)
             },
             receiveValue: { [weak self] result in
                 switch result {
